add and fix tests
