+++ frontend/src/app/api/EndpointConfigs.ts
@@ -1,0 +1,23 @@
+import { ApiMethod } from "src/app/api/FetchHelpers";
+import { environment } from "src/constants/environments";
+
+export interface EndpointConfig {
+  basePath: string;
+  version: string;
+  namespace: string;
+  method: ApiMethod;
+}
+
+export const opportunitySearchEndpoint = {
+  basePath: environment.API_URL,
+  version: "v1",
+  namespace: "opportunities/search",
+  method: "POST" as ApiMethod,
+};
+
+export const fetchOpportunityEndpoint = {
+  basePath: environment.API_URL,
+  version: "v1",
+  namespace: "opportunities",
+  method: "GET" as ApiMethod,
+};
