+++ frontend/tests/pages/search/page.test.tsx
@@ -53,6 +53,7 @@ jest.mock("next/navigation", () => ({
 jest.mock("react", () => ({
   ...jest.requireActual<typeof import("react")>("react"),
   Suspense: ({ fallback }: { fallback: React.Component }) => fallback,
+  cache: (fn: unknown) => fn,
 }));
 
 const fetchMock = jest.fn().mockResolvedValue({
